Remove haproxy_monitor in favour of prometheus
