+++ scripts/monitoring/monitor.py
@@ -10,11 +10,10 @@ import time
 
 import httpx
 
-from scripts.monitoring.haproxy_monitor import GraphiteEvent
 from scripts.monitoring.solr_updater_monitor import get_solr_updater_lag_event
 from scripts.monitoring.utils import (
+    GraphiteEvent,
     bash_run,
-    get_service_ip,
     limit_server,
 )
 from scripts.utils.scheduler import OlAsyncIOScheduler
@@ -59,26 +58,6 @@ def monitor_nginx_logs():
     bash_run(
         f"log_top_ip_counts stats.{bucket}.top_ips",
         sources=["../obfi.sh", "utils.sh"],
-    )
-
-
-@limit_server(["ol-www0"], scheduler)
-@scheduler.scheduled_job('interval', seconds=60)
-async def monitor_haproxy():
-    # Note this is a long-running job that does its own scheduling.
-    # But by having it on a 60s interval, we ensure it restarts if it fails.
-    from scripts.monitoring.haproxy_monitor import main
-
-    web_haproxy_ip = get_service_ip("web_haproxy")
-
-    await main(
-        haproxy_url=f'http://{web_haproxy_ip}:7072/admin?stats',
-        graphite_address=GRAPHITE_URL,
-        prefix='stats.ol.haproxy',
-        dry_run=False,
-        fetch_freq=10,
-        commit_freq=30,
-        agg=None,  # No aggregation
     )
