+++ scripts/monitoring/requirements.txt
@@ -1,3 +1,2 @@
 APScheduler==3.11.0
 py-spy==0.4.0
-requests==2.32.4
